+++ backend-api/internal/query/rest/mapper/product_detail_mapper.go
@@ -57,7 +57,7 @@ func (m *ProductDetailMapper) ToProductResponse(p *models.Product) openapi.Produ
 	sku := p.Sku
 
 	return openapi.Product{
-		Id:            int64(p.ID),
+		Id:            p.ID,
 		Name:          p.Name,
 		Description:   description,
 		Sku:           &sku,
@@ -66,7 +66,7 @@ func (m *ProductDetailMapper) ToProductResponse(p *models.Product) openapi.Produ
 		ImageUrl:      imageURL,
 		InStock:       inStock,
 		StockQuantity: quantity,
-		CategoryId:    int64(p.CategoryID),
+		CategoryId:    p.CategoryID,
 		CategoryName:  categoryName,
 		IsNew:         &isNew,
 		IsFeatured:    &isFeatured,
